+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-03-xx
 
+## [0.64.0] - 2025-03-24
+
 ### Added
 
   - [ET-736] Human Verification